fix(mcp,lsp): three interface defects found by driving every surface (Refs PMAT-199)
Installed 1.12.0 from crates.io and drove every CLI, MCP and LSP interface
against a real project, comparing each tool to its CLI counterpart. None of
these was visible from a schema, from `tools/list`, or from a handler test that
only asserted "returns Ok".

**MCP `forjar_plan` reported every resource as a pending change.**
`ExecutionPlan::changes` carries EVERY resource with its action, NoOp included;
`cli::plan` filters those before counting (plan.rs:262) and the MCP handler did
not. On a fully converged project the CLI said `0 to change` and MCP said 6
changes — so an agent driving forjar through MCP would see a converged system as
needing a full re-apply. It also included phony resources, which are goal-only.

**MCP `forjar_status` returned no machines, ever.** It scanned the state
directory for files with a `.json` extension. A machine's state is a DIRECTORY,
`state/<machine>/state.lock.yaml` — forjar has never written a `.json` lock. The
CLI printed `Machine: local (localhost)` for the same project; MCP returned
`{"machines": []}`. Its existing test wrote `local.json` and commented
"StatusHandler reads .json files from state_dir", codifying the defect. That is
the fifth test this release found asserting a bug as intended behaviour.

**`forjar lsp` was an unrecognized subcommand.** The language server is complete
— stdio framing, diagnostics from the same validator `forjar validate` uses,
completion, 80 passing tests — but no `Commands` variant dispatched it, so no
editor could ever start it. Wired and documented; the doc-parity test required
the book entry. Verified by driving it: `initialize` returns capabilities and
`textDocument/completion` responds.

The phony filter is shared through one named re-export rather than
reimplemented, so `cli::plan` and the MCP handler cannot drift apart again.
`handlers.rs` was split at the config/state boundary to stay under the file
limit.

Verified after the fix, same project: CLI `0 to change` / MCP `0 changes`;
`rm build/app` -> CLI `1 to change` / MCP `1 change (build-app UPDATE)`; CLI
`Machine: local` / MCP `{"machines":[{"name":"local","resource_count":5}]}`.

12570 tests pass.
