+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "forjar"
-version = "1.12.0"
+version = "1.12.1"
 edition.workspace = true
 rust-version = "1.89.0"
 authors = ["Pragmatic AI Labs"]
