+++ src/cli/dispatch.rs
@@ -68,6 +68,10 @@ pub fn dispatch(cmd: Commands, verbose: u8, no_color: bool) -> Result<(), String
         cmd @ Commands::Apply(..) => dispatch_apply_cmd(cmd, verbose),
         Commands::Make(args) => super::make::cmd_make(&args, verbose),
         Commands::ImportMakefile(args) => super::makefile_cmd::cmd_import_makefile(&args),
+        // FJ-2730: the LSP server had 80 passing tests, a complete stdio
+        // implementation and NO entry point — `forjar lsp` was an unrecognized
+        // subcommand. Found by sweeping every interface.
+        Commands::Lsp => super::lsp::cmd_lsp(),
         Commands::Drift(DriftArgs {
             file,
             machine,
