+++ src/mcp/mod.rs
@@ -5,6 +5,7 @@
 //! O(1) dispatch and pforge McpServer for protocol handling.
 
 pub mod handlers;
+pub mod handlers_state;
 pub mod registry;
 pub mod types;
 
@@ -12,6 +13,8 @@ pub mod types;
 mod tests_handlers;
 #[cfg(test)]
 mod tests_handlers_more;
+#[cfg(test)]
+mod tests_parity;
 #[cfg(test)]
 mod tests_registry;
